Separated CMake architecture detection and added ODROID sound support
Modified CMake architecture detection so that it separated GLES
ability detection from the actual VERO4K, RPI, ODROID architecture
detection. This should hopeully make it easier to add different
all-in-one boards such as those listed above in the future.

The above bit was done to enable the ultimate goal of this commit,
which was to add ODROID specific sound options to the GUI, so that
ODROID users could easily select their devices.